+++ CHANGELOG.rst
@@ -1,7 +1,19 @@
 TransitionMatrix Change Log
 ===========================
 
-v0.1.0
-==========
+v0.1.0 (11-11-2017)
+===================
 
-* First release of the package (11-11-2017)
+* First release of the package
+
+v0.1.1 (03-12-2017)
+===================
+
+* TransitionMatrix model: new methods to merge States, fix problematic probability matrices, I/O API's
+* TransitionMatrixSet mode: json and csv readers, methods for set-wise manipulations
+* Additional multiperiod datasets (Standard and Poors historical corporating rating transition rates)
+* Enhanced matrix comparison functionality
+* Three additional example workflows
+    * fixing multiperiod matrices (completing State Space)
+    * adjusting matrices for withdrawn entries
+    * generating full  multi-period sets from limited observations